+++ threedi_tools.py
@@ -21,6 +21,7 @@
  ***************************************************************************/
 """
 
+import os
 import os.path
 
 from PyQt4.QtCore import (QSettings, QTranslator, qVersion, QCoreApplication,
@@ -32,13 +33,14 @@ from qgis.core import QgsMapLayerRegistry
 import resources  # NoQa
 
 # Import the code of the tools
+from .stats.utils import STATS_CACHE_LAYER_POSTFIX
 from .threedi_result_selection import ThreeDiResultSelection
 from .threedi_toolbox import ThreeDiToolbox
 from .threedi_graph import ThreeDiGraph
 from .threedi_sideview import ThreeDiSideView
 from .views.timeslider import TimesliderWidget
 from .views.map_animator import MapAnimator
-from .utils.user_messages import pop_up_info, log
+from .utils.user_messages import pop_up_info, pop_up_question, log
 from .models.datasources import TimeseriesDatasourceModel
 from .utils.qprojects import ProjectStateMixin
 from .utils.layer_tree_manager import LayerTreeManager
@@ -64,6 +66,44 @@ class About(object):
 
         pop_up_info("3Di Tools versie %s" % version,
                     "About", self.iface.mainWindow())
+
+    def on_unload(self):
+        pass
+
+
+class CacheClearer(object):
+    def __init__(self, iface, parent):
+        self.iface = iface
+        self.icon_path = ':/plugins/ThreeDiToolbox/icon_broom.png'
+        self.menu_text = "Clear cache"
+        self.parent = parent
+
+    def run(self):
+        spatialite_filepaths = [
+            item.spatialite_cache_filepath() for
+            item in self.parent.ts_datasource.rows if
+            os.path.exists(item.spatialite_cache_filepath())]
+        result_dirs = [os.path.dirname(item.file_path.value) for
+                       item in self.parent.ts_datasource.rows]
+        csv_filepaths = [
+            os.path.join(_dir, f) for _dir in result_dirs
+            for f in os.listdir(_dir)
+            if f.endswith(STATS_CACHE_LAYER_POSTFIX)]
+        cached = set(spatialite_filepaths + csv_filepaths)
+        if not cached:
+            pop_up_info("No cached files found.")
+            return
+
+        yes = pop_up_question(
+            "The following files will be deleted:\n" +
+            ',\n'.join(cached) +
+            "\n\nContinue?")
+
+        if yes:
+            for f in cached:
+                os.remove(f)
+            pop_up_info("Cache cleared. You may need to restart QGIS and "
+                        "reload your data.")
 
     def on_unload(self):
         pass
@@ -130,6 +170,7 @@ class ThreeDiTools(QObject, ProjectStateMixin):
 
         self.tools = [
             About(iface),
+            CacheClearer(iface, self),
             ThreeDiResultSelection(iface, self.ts_datasource),
             ThreeDiToolbox(iface, self.ts_datasource),
             self.graph_tool,